add all soap operations for album + refactoring

+++ dbconnector/src/main/java/ru/rti/kettu/sbjava/controller/AlbumController.java
@@ -10,6 +10,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import static org.springframework.util.ObjectUtils.isEmpty;
+import static ru.rti.kettu.sbjava.mapping.http.AlbumHttpMapping.mapHttpRequest;
 
 @RestController
 @RequestMapping(path = "/album")
@@ -21,18 +22,21 @@ public class AlbumController {
         this.musicService = musicService;
     }
 
-    @GetMapping (path = "/createAlbum")
-    public String createAlbum (String author, String name, int year) {
-        return "New album was created with id: " + musicService.createAlbumInfo(author, name, year);
+    @GetMapping(path = "/createAlbum")
+    public String createAlbum(String author, String name, int year) {
+        Album album = mapHttpRequest(null, author, name, year);
+        return "New album was created with id: " + musicService.createAlbumInfo(album);
     }
 
-    @GetMapping (path = "/deleteAlbum")
-    public void deleteAlbum (Long id) {
-        musicService.deleteAlbumInfo(id);
+    @GetMapping(path = "/deleteAlbum")
+    public boolean deleteAlbum(Long id) {
+        if (isEmpty(id)) {
+            return musicService.deleteAlbumInfo(id);
+        } else return musicService.deleteAllAlbumInfo();
     }
 
-    @GetMapping (path = "/getAlbum")
-    public List<Album> getAllAlbums (Long id) {
+    @GetMapping(path = "/getAlbum")
+    public List<Album> getAllAlbums(Long id) {
         if (isEmpty(id))
             return musicService.getAllAlbums();
         else return Arrays.asList(musicService.getAlbumById(id));
@@ -40,6 +44,7 @@ public class AlbumController {
 
     @GetMapping(path = "/updateAlbum")
     public Album updateAlbum(Long id, String author, String name, int year) {
-        return musicService.updateAlbum(id, name, author, year);
+        Album album = mapHttpRequest(id, name, author, year);
+        return musicService.updateAlbum(album);
     }
 }
